+++ src/Point.java
@@ -1,31 +1,31 @@
 import java.io.Serializable;
 
 public class Point implements Serializable{
-    private float posX;
-    private float posY;
+    private int posX;
+    private int posY;
 
-    public Point(float posX , float posY) {
+    public Point(int posX , int posY) {
         this.posX = posX;
         this.posY = posY;
     }
 
-    public float getX() { return posX; }
+    public int getX() { return posX; }
 
-    public float getY() {
+    public int getY() {
         return posY;
     }
 
-    public void setX(float posX) {
+    public void setX(int posX) {
         this.posX = posX;
     }
 
-    public void setY(float posY) {
+    public void setY(int posY) {
         this.posY = posY;
     }
 
-    public float distSquared(Point other) {
-        float distX = this.posX - other.posX;
-        float distY = this.posY - other.posY;
+    public int distSquared(Point other) {
+        int distX = this.posX - other.posX;
+        int distY = this.posY - other.posY;
         return distX*distX + distY*distY;
     }
 }
